+++ CHANGES.rst
@@ -2,7 +2,14 @@ Changelog
 =========
 
 
-1.0a3 (unreleased)
+1.0a4 (unreleased)
+------------------
+
+- Redirect to `HTTP_REFERER` instead context's url after
+  `add/remove/remove item` actions.
+  [gbastien]
+
+1.0a3 (2026-03-20)
 ------------------
 
 - Replaced external_session_link p by span.
@@ -29,6 +36,20 @@ Changelog
   [chris-adam]
 - Added external watchers for esign sessions.
   [chris-adam, sgeulette]
+- Added possibility to have elements of the same context to belong to different sessions.
+  [chris-adam]
+- Manage file added again to same session, data is updated.
+  [gbastien]
+- Configured the `@@remove-from-esign-session` the same way as `@@remove-item-from-esign-session`
+  so relying on an `available` method to show it only if a context is in a session.
+  [gbastien]
+- Fixed faceted viewlet broken because sessions format changed from list to OrderedDict.
+  [gbastien]
+- Turned file info in session annotation from `dict` to `PersistentMapping`.
+  Entire annotation structure is now persistent, remove `_p_changed`.
+  [gbastien]
+- Use `@CachedProperty` for `FacetedSessionInfoViewlet.sessions` and `ItemSessionInfoViewlet.sessions`.
+  [gbastien]
 - Added separated fingerpointing log.
   [chris-adam]